// frontend/api/models/GetCollegeFollowsByUserResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GetCollegeFollowsByUserResponse = {
  /**
   * @description A URL to the JSON Schema for this object.
   * @type string | undefined, uri
   */
  readonly $schema?: string;
  /**
   * @description The colleges the given user follows
   * @type array,null
   */
  college_ids: string[] | null;
  /**
   * @description ID of the user
   * @type string
   */
  user_id: string;
};