// frontend/api/models/GetSportFollowsByUserResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GetSportFollowsByUserResponse = {
  /**
   * @description A URL to the JSON Schema for this object.
   * @type string | undefined, uri
   */
  readonly $schema?: string;
  /**
   * @description The sports the given user follows
   * @type array,null
   */
  sport_ids: string[] | null;
  /**
   * @description ID of the user
   * @type string
   */
  user_id: string;
};