// frontend/api/models/GetTagFollowsByUserResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type GetTagFollowsByUserResponse = {
  /**
   * @description A URL to the JSON Schema for this object.
   * @type string | undefined, uri
   */
  readonly $schema?: string;
  /**
   * @description The tags the given user follows
   * @type array,null
   */
  tag_ids: string[] | null;
  /**
   * @description ID of the user
   * @type string
   */
  user_id: string;
};