// frontend/api/models/GetTagResponse.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type GetTagResponse = {
    /**
     * @description A URL to the JSON Schema for this object.
     * @type string | undefined, uri
    */
    readonly $schema?: string;
    /**
     * @description ID of the tag
     * @type string
    */
    id: string;
    /**
     * @description The name of the tag
     * @type string
    */
    name: string;
};