// frontend/api/models/SportResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type SportResponse = {
  /**
   * @description A URL to the JSON Schema for this object.
   * @type string | undefined, uri
   */
  readonly $schema?: string;
  /**
   * @description ID of the sport
   * @type string
   */
  id: string;
  /**
   * @description Name of the sport
   * @type string
   */
  name: string;
  /**
   * @description Number of players
   * @type integer | undefined, int32
   */
  popularity?: number;
};